Create 5.Largest value in each level.cpp
Given a binary tree, find the largest value in each level.

Your task :
You don't need to read input or print anything. Your task is to complete the function largestValues() which takes the root node of the tree as input and returns a vector containing the largest value in each level.