Create 16.Merge_K_sorted_Linked_lists.cpp
Given K sorted linked lists of different sizes. The task is to merge them in such a way that after merging they will be a single sorted linked list.

The task is to complete the function mergeKList() which merges the K given lists into a sorted one. The printing is done automatically by the driver code.The task is to complete the function mergeKList() which merges the K given lists into a sorted one. The printing is done automatically by the driver code.